README.md
Looking for the best WEEX referral code in 2025? Use the exclusive promo code: 3AND when you sign up on WEEX to unlock a new user bonus of up to 30,000 USDT, along with a 50% discount on trading fees. Whether you're a beginner or an experienced crypto trader, this WEEX signup bonus makes joining WEEX even more rewarding.